Add files via upload
In the present context of AIUB there is no bank account system for student. 
A student bank account management system refers to the system that 
allows students to make money transfer, withdraw and to deposit their 
money. If this system is added in AIUB then student do not need to go 
outside of the University to deposit or withdraw their money. They can 
easily make their payment of the university by using this system.   
Our main focus on AIUB mobile banking system policy that will be 
practiced by student of AIUB. It is one type of banking system towards 
university going students to deposit or withdraw money without any charge. 
A banking system has Saving, current, fixed deposit, joint, and minor 
deposit accounts. Any student who wishes to become a new bank customer 
must pick from one of these options. A new student to the bank can create 
an individual account, a joint account, or a minor account, depending on his 
or her preferences. There are no limits on the quantity or amount of deposits 
for cash transactions over 49,999 TK. When a student with a savings bank 
account makes a cash deposit or cash withdrawal, the bank authority 
responsible for this section will update the bank database and the customer's 
account database by entering all transaction details (customer name, 
account number, phone number, deposited/withdrawal amount, and PAN 
number if the transaction exceeds 49,999 TK).  It also offers fund transfer 
services, which allow money to be transferred from one customer's account 
to another. Money can be transferred inside the same bank; however, 
transferring money to the account holder of another bank incurs no fee.    
Students can get their money through ATM card, an internet website or 
mobile device. By using the ATM card, a customer can withdraw cash from 
either checking or savings account, query the balance of an account, or 
transfer funds from one account to another.  A transaction is initiated when 
a customer inserts an ATM card into the card reader. Encoded on the 
magnetic strip on the back of the ATM card is the card number, the start 
date, and the expiration date. Manually managing all of these bank data and 
documents is impossible. As a result, there is a need for software that can 
manage all of the details on both the university students' and bank officials' 
sides. The software's primary goal is to automate all aspects of bank 
operations, both for students and for bank officials. As a result, the banking 
management system is established to address this problem. The system 
allows students to create accounts, deposit/withdraw money from their 
accounts. This system will help to do student’s banking activities such as 
deposits and withdrawals without having to locate an open location and 
without any charge.